requirements.txt
streamlit
psycopg2-binary
sqlalchemy
python-dotenv
pandas
plotly